+++ CHANGES.md
@@ -1,5 +1,9 @@
 ## dev
 
+* Space media with blank `p`: easier focus on mobile
+* Skip empty blocks with `getContent` (#204)
+* Focus on load
+
 ## 0.12.3 - 2016-05-26
 
 * Add label to Author & Publisher dropdown (#203)